Split test_validate.cpp
